chore(docs): rename `lib` to `src`
