+++ src/detectCurrentPlatform.js
