+++ src/platforms.js
