+++ test/utils.js
@@ -3,7 +3,7 @@ var testSetup = require('redtape');
 var temp = require('temp');
 var fs   = require('fs');
 var path = require('path');
-var utils = require('./../lib/utils');
+var utils = require('./../src/utils');
 var DecompressZip = require('decompress-zip');
 var _ = require('lodash');
 var EventEmitter = require('events').EventEmitter;
